customer_churn_prediction/requirements.txt
pandas
numpy
scikit-learn
joblib
mlflow
fastapi
uvicorn
python-dotenv
pydantic
